requirements.txt
Flask==2.0.2
gunicorn
azure-eventhub
azure-eventhub-checkpointstoreblob-aio
azure-identity
aiohttp
